Refactoring of the code base. All methods should work as before. Widgets are now in separate lua files
