+++ fun.lua
@@ -4,9 +4,13 @@
 --
 -----------------------------------------------------------------------------------------
 
+-- corona
 local composer = require( "composer" )
-local mui = require( "materialui.mui" )
 local widget = require( "widget" )
+
+-- mui
+local muiData = require( "mui-data" )
+local mui = require( "materialui.mui" )
 
 local scene = composer.newScene()
