+++ materialui/mui-example.lua
@@ -1,0 +1,21 @@
+--
+-- mui-example template module, extend mui
+--
+
+-- mui
+local muiData = require( "mui-data" )
+
+local mathFloor = math.floor
+local mathMod = math.fmod
+local mathABS = math.abs
+
+local M = muiData.M -- {} -- for module array/table
+
+-- define methods here
+--[[--
+function M.exampleMethod()
+
+end
+--]]--
+
+return M
