+++ mui-data.lua
@@ -1,0 +1,3 @@
+--my global space
+local M = {}
+return M
